refactor(simulation): move history flow into feature module
