+++ src/main/java/com/renewsim/backend/simulation_service/history/application/SimulationHistoryRowResult.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.application.historySimulation;
+package com.renewsim.backend.simulation_service.history.application;
 
 public record SimulationHistoryRowResult(
         String id,
