+++ src/main/java/com/renewsim/backend/simulation_service/history/web/SimulationHistoryController.java
@@ -1,7 +1,7 @@
 package com.renewsim.backend.simulation_service.history.web;
 
 import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteRealSimulationUseCase;
-import com.renewsim.backend.simulation_service.application.historySimulation.ListUserRealSimulationsUseCase;
+import com.renewsim.backend.simulation_service.history.application.ListUserRealSimulationsUseCase;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContext;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContextFactory;
 import com.renewsim.backend.simulation_service.history.web.dto.ListUserSimulationsResponseDTO;
