+++ src/test/java/com/renewsim/backend/simulation_service/create/application/CreateSimulationServiceTest.java
@@ -10,8 +10,8 @@ import com.renewsim.backend.simulation_service.application.dashboard.GetPortfoli
 import com.renewsim.backend.simulation_service.application.dashboard.PortfolioDashboardRiskAlert;
 import com.renewsim.backend.simulation_service.application.dashboard.PortfolioDashboardResult;
 import com.renewsim.backend.simulation_service.detail.application.GetSimulationService;
-import com.renewsim.backend.simulation_service.application.historySimulation.ListSimulationsService;
-import com.renewsim.backend.simulation_service.application.historySimulation.UserSimulationListResult;
+import com.renewsim.backend.simulation_service.history.application.ListSimulationsService;
+import com.renewsim.backend.simulation_service.history.application.UserSimulationListResult;
 import com.renewsim.backend.simulation_service.application.port.out.PvgisSolarResourcePort;
 import com.renewsim.backend.simulation_service.application.port.out.SimulationRecordRepositoryPort;
 import com.renewsim.backend.simulation_service.application.port.out.TechnologyLookupPort;
