+++ src/test/java/com/renewsim/backend/simulation_service/history/web/SimulationHistoryControllerTest.java
@@ -3,9 +3,9 @@ package com.renewsim.backend.simulation_service.history.web;
 import com.renewsim.backend.auth_service.domain.AuthenticatedUser;
 import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteRealSimulationUseCase;
 import com.renewsim.backend.simulation_service.detail.application.GetRealSimulationUseCase;
-import com.renewsim.backend.simulation_service.application.historySimulation.ListUserRealSimulationsUseCase;
-import com.renewsim.backend.simulation_service.application.historySimulation.SimulationHistoryRowResult;
-import com.renewsim.backend.simulation_service.application.historySimulation.UserSimulationListResult;
+import com.renewsim.backend.simulation_service.history.application.ListUserRealSimulationsUseCase;
+import com.renewsim.backend.simulation_service.history.application.SimulationHistoryRowResult;
+import com.renewsim.backend.simulation_service.history.application.UserSimulationListResult;
 import com.renewsim.backend.simulation_service.detail.web.SimulationDetailController;
 import com.renewsim.backend.simulation_service.history.web.dto.ListUserSimulationsResponseDTO;
 import org.junit.jupiter.api.DisplayName;
